# Conflicts:
#	bundles/AdminBundle/Controller/Admin/ElementController.php
#	bundles/AdminBundle/Controller/Admin/SettingsController.php
#	bundles/AdminBundle/Resources/public/js/pimcore/asset/helpers/gridConfigDialog.js
#	bundles/AdminBundle/Resources/public/js/pimcore/document/tags/image.js
#	bundles/AdminBundle/Resources/public/js/pimcore/object/classificationstore/keySelectionWindow.js
#	bundles/AdminBundle/Resources/public/js/pimcore/settings/metadata/predefined.js

+++ bundles/AdminBundle/Controller/Admin/SettingsController.php
@@ -132,7 +132,9 @@ class SettingsController extends AdminController
     }
 
     /**
-     * @Route("/metadata", name="pimcore_admin_settings_metadata", methods={"POST"})
+     * Used by the predefined metadata grid
+     *
+     * @Route("/predefined-metadata", name="pimcore_admin_settings_metadata", methods={"POST"})
      *
      * @param Request $request
      *
@@ -1333,6 +1335,7 @@ class SettingsController extends AdminController
         $pipe = Asset\Image\Thumbnail\Config::getByName($request->get('name'));
         $settingsData = $this->decodeJson($request->get('settings'));
         $mediaData = $this->decodeJson($request->get('medias'));
+        $mediaOrder = $this->decodeJson($request->get('mediaOrder'));
 
         foreach ($settingsData as $key => $value) {
             $setter = 'set' . ucfirst($key);
@@ -1342,6 +1345,14 @@ class SettingsController extends AdminController
         }
 
         $pipe->resetItems();
+
+        uksort($mediaData, function ($a, $b) use ($mediaOrder) {
+            if ($a === 'default') {
+                return -1;
+            }
+
+            return ($mediaOrder[$a] < $mediaOrder[$b]) ? -1 : 1;
+        });
 
         foreach ($mediaData as $mediaName => $items) {
             foreach ($items as $item) {
@@ -1755,34 +1766,33 @@ class SettingsController extends AdminController
             if ($request->get('xaction') == 'destroy') {
                 $id = $data['id'];
                 $setting = WebsiteSetting::getById($id);
-                $setting->delete();
+                if ($setting instanceof WebsiteSetting) {
+                    $setting->delete();
 
-                return $this->adminJson(['success' => true, 'data' => []]);
+                    return $this->adminJson(['success' => true, 'data' => []]);
+                }
             } elseif ($request->get('xaction') == 'update') {
                 // save routes
                 $setting = WebsiteSetting::getById($data['id']);
-
-                switch ($setting->getType()) {
-                    case 'document':
-                    case 'asset':
-                    case 'object':
-                        if (isset($data['data'])) {
-                            $path = $data['data'];
-                            $element = null;
-                            if ($path != null) {
-                                $element = Element\Service::getElementByPath($setting->getType(), $path);
+                if ($setting instanceof WebsiteSetting) {
+                    switch ($setting->getType()) {
+                        case 'document':
+                        case 'asset':
+                        case 'object':
+                            if (isset($data['data'])) {
+                                $element = Element\Service::getElementByPath($setting->getType(), $data['data']);
+                                $data['data'] = $element;
                             }
-                            $data['data'] = $element;
-                        }
-                        break;
+                            break;
+                    }
+
+                    $setting->setValues($data);
+                    $setting->save();
+
+                    $data = $this->getWebsiteSettingForEditMode($setting);
+
+                    return $this->adminJson(['data' => $data, 'success' => true]);
                 }
-
-                $setting->setValues($data);
-                $setting->save();
-
-                $data = $this->getWebsiteSettingForEditMode($setting);
-
-                return $this->adminJson(['data' => $data, 'success' => true]);
             } elseif ($request->get('xaction') == 'create') {
                 unset($data['id']);
